Something is upsetting fps

+++ src/game/clock.cpp
@@ -1,7 +1,7 @@
 #include "game/clock.h"
 
 void Clock::tick() {
-    uint32_t ticks = SDL_GetTicks();
+    Uint64 ticks = SDL_GetTicks();
     delta = ticks - previousTicks;
     previousTicks = ticks;
     fps = (delta > 0) ? 1000.0f / delta : 0.0f;
