+++ src/game/clock.h
@@ -9,11 +9,11 @@
  */
 class Clock {
     private:
-        uint32_t previousTicks = 0;
+        Uint64 previousTicks = 0;
 
     public:
-        uint32_t delta = 0;    ///< Time difference in milliseconds between the current and the previous tick.
-        uint32_t fps = 0;      ///< The number of frames per second.
+        Uint64 delta = 0;    ///< Time difference in milliseconds between the current and the previous tick.
+        Uint64 fps = 0;      ///< The number of frames per second.
 
         /**
          * @brief Updates delta and fps values based on the current tick.
